[fs] support hfs.ls on a bucket (#14176)
Teaches `hfs.ls('gs://bucket/')` to list the files and directories at
the top-level of the bucket.

In `main` that command raises because this line of `_ls_no_glob` raises:

```python3
maybe_sb_and_t, maybe_contents = await asyncio.gather(
    self._size_bytes_and_time_modified_or_none(path), ls_as_dir()
)
```

In particular, `statfile` raises a cloud-specific, esoteric error about
a malformed URL or empty object names:

```python3
async def _size_bytes_and_time_modified_or_none(self, path: str) -> Optional[Tuple[int, float]]:
    try:
        # Hadoop semantics: creation time is used if the object has no notion of last modification time.
        file_status = await self.afs.statfile(path)
        return (await file_status.size(), file_status.time_modified().timestamp())
    except FileNotFoundError:
        return None
```

I decided to add a sub-class of `FileNotFoundError` which is
self-describing: `IsABucketError`.

I changed most methods to raise that error when given a bucket URL. The
two interesting cases:

1. `isdir`. This raises an error but I could also see this returning
`True`. A bucket is like a directory whose path/name is empty.

2. `isfile`. This returns False but I could also see this raising an
error. This just seems convenient, we know the bucket is not a file so
we should say so.

---

Apparently `hfs.ls` had no current tests because the globbing system
doesn't work with Azure https:// URLs. I fixed it to use
`AsyncFSURL.with_new_path_component` which is resilient to Azure https
weirdness. However, I had to change `with_new_path_component` to treat
an empty path in a special way. I wanted this to hold:

```
actual = str(afs.parse_url('gs://bucket').with_new_path_component('bar'))
expected = 'gs://bucket/bar'
assert actual == expected
```

But `with_new_path_component` interacts badly with
`GoogleAsyncFSURL.__str__` to return this:

```
'gs://bucket//bar'
```

+++ hail/python/hailtop/aiocloud/aiogoogle/client/storage_client.py
@@ -21,6 +21,7 @@ from hailtop.aiotools.fs import (
     FileAndDirectoryError,
     MultiPartCreate,
     UnexpectedEOFError,
+    IsABucketError,
 )
 from hailtop.aiotools import FeedableAsyncIterable, WriteBuffer
 
@@ -578,6 +579,9 @@ class GoogleStorageAsyncFSURL(AsyncFSURL):
         self._bucket = bucket
         self._path = path
 
+    def __repr__(self):
+        return f'GoogleStorageAsyncFSURL({self._bucket}, {self._path})'
+
     @property
     def bucket_parts(self) -> List[str]:
         return [self._bucket]
@@ -596,6 +600,9 @@ class GoogleStorageAsyncFSURL(AsyncFSURL):
 
     def with_path(self, path) -> 'GoogleStorageAsyncFSURL':
         return GoogleStorageAsyncFSURL(self._bucket, path)
+
+    def with_root_path(self) -> 'GoogleStorageAsyncFSURL':
+        return self.with_path('')
 
     def __str__(self) -> str:
         return f'gs://{self._bucket}/{self._path}'
@@ -645,8 +652,11 @@ class GoogleStorageAsyncFS(AsyncFS):
         return url.startswith('gs://')
 
     @staticmethod
-    def parse_url(url: str) -> GoogleStorageAsyncFSURL:
-        return GoogleStorageAsyncFSURL(*GoogleStorageAsyncFS.get_bucket_and_name(url))
+    def parse_url(url: str, *, error_if_bucket: bool = False) -> GoogleStorageAsyncFSURL:
+        fsurl = GoogleStorageAsyncFSURL(*GoogleStorageAsyncFS.get_bucket_and_name(url))
+        if error_if_bucket and fsurl._path == '':
+            raise IsABucketError
+        return fsurl
 
     @staticmethod
     def get_bucket_and_name(url: str) -> Tuple[str, str]:
@@ -673,25 +683,26 @@ class GoogleStorageAsyncFS(AsyncFS):
         return (bucket, name)
 
     async def open(self, url: str) -> GetObjectStream:
-        bucket, name = self.get_bucket_and_name(url)
-        return await self._storage_client.get_object(bucket, name)
+        fsurl = self.parse_url(url, error_if_bucket=True)
+        return await self._storage_client.get_object(fsurl._bucket, fsurl._path)
 
     async def _open_from(self, url: str, start: int, *, length: Optional[int] = None) -> GetObjectStream:
-        bucket, name = self.get_bucket_and_name(url)
+        fsurl = self.parse_url(url, error_if_bucket=True)
         range_str = f'bytes={start}-'
         if length is not None:
             assert length >= 1
             range_str += str(start + length - 1)
-        return await self._storage_client.get_object(bucket, name, headers={'Range': range_str})
+        return await self._storage_client.get_object(fsurl._bucket, fsurl._path, headers={'Range': range_str})
 
     async def create(self, url: str, *, retry_writes: bool = True) -> WritableStream:
-        bucket, name = self.get_bucket_and_name(url)
+        fsurl = self.parse_url(url, error_if_bucket=True)
         params = {'uploadType': 'resumable' if retry_writes else 'media'}
-        return await self._storage_client.insert_object(bucket, name, params=params)
+        return await self._storage_client.insert_object(fsurl._bucket, fsurl._path, params=params)
 
     async def multi_part_create(
         self, sema: asyncio.Semaphore, url: str, num_parts: int
     ) -> GoogleStorageMultiPartCreate:
+        self.parse_url(url, error_if_bucket=True)
         return GoogleStorageMultiPartCreate(sema, self, url, num_parts)
 
     async def staturl(self, url: str) -> str:
@@ -705,8 +716,8 @@ class GoogleStorageAsyncFS(AsyncFS):
 
     async def statfile(self, url: str) -> GetObjectFileStatus:
         try:
-            bucket, name = self.get_bucket_and_name(url)
-            return GetObjectFileStatus(await self._storage_client.get_object_metadata(bucket, name), url)
+            fsurl = self.parse_url(url, error_if_bucket=True)
+            return GetObjectFileStatus(await self._storage_client.get_object_metadata(fsurl._bucket, fsurl._path), url)
         except aiohttp.ClientResponseError as e:
             if e.status == 404:
                 raise FileNotFoundError(url) from e
@@ -784,12 +795,12 @@ class GoogleStorageAsyncFS(AsyncFS):
 
     async def isfile(self, url: str) -> bool:
         try:
-            bucket, name = self.get_bucket_and_name(url)
+            fsurl = self.parse_url(url)
             # if name is empty, get_object_metadata behaves like list objects
             # the urls are the same modulo the object name
-            if not name:
+            if not fsurl._path:
                 return False
-            await self._storage_client.get_object_metadata(bucket, name)
+            await self._storage_client.get_object_metadata(fsurl._bucket, fsurl._path)
             return True
         except aiohttp.ClientResponseError as e:
             if e.status == 404:
@@ -797,10 +808,10 @@ class GoogleStorageAsyncFS(AsyncFS):
             raise
 
     async def isdir(self, url: str) -> bool:
-        bucket, name = self.get_bucket_and_name(url)
-        assert not name or name.endswith('/'), name
-        params = {'prefix': name, 'delimiter': '/', 'includeTrailingDelimiter': 'true', 'maxResults': 1}
-        async for page in await self._storage_client.list_objects(bucket, params=params):
+        fsurl = self.parse_url(url, error_if_bucket=True)
+        assert not fsurl._path or fsurl.path.endswith('/'), fsurl._path
+        params = {'prefix': fsurl._path, 'delimiter': '/', 'includeTrailingDelimiter': 'true', 'maxResults': 1}
+        async for page in await self._storage_client.list_objects(fsurl._bucket, params=params):
             prefixes = page.get('prefixes')
             items = page.get('items')
             return bool(prefixes or items)
@@ -808,6 +819,8 @@ class GoogleStorageAsyncFS(AsyncFS):
 
     async def remove(self, url: str) -> None:
         bucket, name = self.get_bucket_and_name(url)
+        if name == '':
+            raise IsABucketError(url)
         try:
             await self._storage_client.delete_object(bucket, name)
         except aiohttp.ClientResponseError as e:
